Update documentation; add helper fns to primitive struct; UTs
